Implement MM algorithm for metric projection.
This required a change to the penalty function interface.
The algorithm maps involve a resolvent-like R = (I/rho + L)^-1 operator
based on a linear operator L. In general, though, the matrix I may be
replaced by a diagonal matrix W. In this case, it is better to have an
interface that accepts the W matrix so that the linear operator can be
updated in linear time with respect to the matrix dimensions.

Note: The MM algorithm has unusual memory allocations. Probably some
type instability.