Detected and fixed catastrophic cancellation
The code for the sparsity projection contained a small roundoff error
that turns out to be catastrophic in the clustering algorithm.
Specifically, we knew that `y1 = Diagonal(W) * D * vec(X)` is different
from `cvxclst_apply_fusion_matrix(y2, W, X)` but approximately equal.
This difference seemed to manifest in cancellation when computing the
vector `y - p`, where `p` is the projection of `y`. Further testing is
needed to confirm this theory but it seems to be the only explanation.

We now use Julia's `sortperm` to identify large blocks in distances
between centroid as it turns out to (i) eliminate the artifacts from
before and (ii) runs faster than my approach that fuses sorting and
distance computations.