Tweak algorithm parameters and annealing schedule
The penalty function is now set to update every 250 iterations
using a multiplier of 2.0. This choice reaches a peak penalty
coefficient on the order 10^6 and appears to mitigate numerical
issues. I settled on this choice by checking the rate of decay
in the penalty term, rho * dist^2.

My choice appears to guarantee convergence in all my benchmarks.
This is particularly important for the clustering problems because
our heuristic is sensitive to the quality of solutions.