Major overhaul of internal interface to make ADMM
This commit provides a common solution interface made possible by the
LinearMaps package. As a first step, we implement a steepest descent
algorithm for metric projection. There is a lot of reorganization:

* common.jl: Provides the standardized `optimize!` function to solve any
optimization problem with the proximal distance approach. It makes use
of `NamedTuple` to simplify function signatures without creating lots
of types specific to permutation of (problem, algorithm).

* linear_operators.jl -> operators.jl: This file should implement the
internal FusionMatrix interface for a specific problem. It should also
include any required projection operators.

* <algorith_name>.jl: This file should provide functions to evaluate the
objective for a problem, along with any gradients and projections. It
should also provide an interface to a solver for a specific problem by
setting up data structures and calling `optimize!`.

* penalty.jl: No longer needed.

* acceleration.jl: Simplified the interface thanks to `trivec_view`.