Implement `Super Memory Gradient` MM subspace method (thanks to Oscar)
- Adds a new `AlgorithmOption`, `MMSubSpace{K}` where `K` is the number
of gradients to use in the subspace.
- Implements the basic algorithm map for `metric_projection`.
- Introduces a branch in `optimize!` to handle subspace updates.

The basic strategy is to solve the linesearch problem using either
LSQR or CG. Setting up each linear system works exactly as with MM, but
with slightly different LHS and RHS blocks. The current implementation
makes use of `BlockMap`s from LinearMaps.jl, which necessarily create
temporary arrays due to block-by-block evaluation. We will need to a
new `LinearMap` type to avoid allocations in larger problems.